+++ src/views/browser_view.cpp
@@ -204,6 +204,9 @@ void BrowserView::show_job_dialog(rclone::JobType type) {
 
     m_job_dialog = std::make_unique<JobEditDialog>(type, src, dst, includes,
         [this](rclone::Job job) { signal_job_created.emit(job); });
+    m_job_dialog->set_save_callback([this](rclone::Job job) {
+        signal_job_saved.emit(job);
+    });
     if (auto* win = dynamic_cast<Gtk::Window*>(get_root()))
         m_job_dialog->set_transient_for(*win);
     m_job_dialog->present();